Seperated resampling/projection	into its own procedure for alternate projections.
hzpfsim_img.pro now requires a parameter to define how the projection onto the FPA is performed. The options are currently 'simple' or '7_fibers'. If 'simple' is selected (which is also the default), the projection of the simple slit (star fiber + cal fiber) is performed by simple_slit_projection.pro. If '7_fibers' is selected, the routine fiber_projection.pro will project the star into 7 sub-fibers and the cal into two fibers on the top and bottom of the stellar fibers. Check the parameters for fiber_projection.pro to see how to change the geometry and flux ratios through the fibers. The generalizability of the fiber arrangement hasn't been tested yet, I've only used it with the initial set of beam separations provided by Chad. Alternative extractions have not yet been implemented, so only the standard hzpfex_basic2.pro is called. This will correctly extract the spectrum if 'simple' is selected for projection, but will not correctly
extract the '7_fibers' projection.